Adjust pass defaults
