+++ backend/pipeline/passes/__init__.py
@@ -1,0 +1,6 @@
+"""Pass execution package."""
+from __future__ import annotations
+
+from .runner import run_all_passes_async
+
+__all__ = ["run_all_passes_async"]
